### 2024-09-11 #657 @spsanderson
Is my TidyDensity package a good fit here?
Submitting Author Name: Steven P. Sanderson II, MPH
Submitting Author Github Handle: @spsanderson
Other Package Authors Github handles: None
Repository:  https://github.com/spsanderson/TidyDensity
Submission type: Pre-submission

---

```
Package: TidyDensity
Title: Functions for Tidy Analysis and Generation of Random Data
Version: 1.5.0.9000
    person("Steven","Sanderson", email = "spsanderson@gmail.com", role = c("aut","cre","cph"),
      comment = c(ORCID = "0009-0006-7661-8247")
      )
    )
    To make it easy to generate random numbers based upon the underlying stats
    distribution functions. All data is returned in a tidy and structured
    format making working with the data simple and straight forward. Given that the
    data is returned in a tidy 'tibble' it lends itself to working with the rest of the
    'tidyverse'.
    rlang (>= 0.4.11),
    plotly,
    actuar,
    patchwork,
    survival,
    nloptr,
    broom,
    tidyselect,
    stringr
    EnvStats
```

- Explain how and why the package falls under these categories (briefly, 1-2 sentences).  Please note any areas you are unsure of: I think this package fits here because it is designed to help people generate distributions and to see how distributions change in shape when their parameters change.

- If submitting a statistical package, have you already [incorporated documentation of standards into your code via the **srr** package](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr)? No I have not used srr package.

-   Who is the target audience and what are scientific applications of this package?  Anyone that wants to see how distributions change with parameters (students) or anyone that is looking for parameter estimation etc.

-   Are there other R packages that accomplish the same thing? If so, how does yours differ or meet [our criteria for best-in-category](https://ropensci.github.io/dev_guide/policies.html#overlap)? There are packages like Distributional, actuar and EnvStats, but to my knowledge nothing out there that has as many parameter estimation tools and nothing that brings back all generators as tidyverse compliant tibbles with r, p, d, and q all set in one function.

-   (If applicable) Does your package comply with our [guidance around _Ethics, Data Privacy and Human Subjects Research_](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)? I don't collect any telemetry so I would say yes.

-  Any other questions or issues we should be aware of?: I don't really know.
