### 2024-09-11 #657 @mpadge
@spsanderson The stats team will discuss your query and get back to you asap. In regard to *Probability Distributions* standards, our [stats dev-guide](https://stats-devguide.ropensci.org/standards.html#standards-distributions) states:

> Unlike most other categories of standards, packages which fit in this category will also generally be expected to fit into at least one other category of statistical software.

This would be first package that goes against that general principle. That said, it is clearly well-developed, and looks like a significant contribution to probability distributions in general. We just have to decide whether we'll relax that condition for your submission.
