### 2024-09-11 #657 @spsanderson
Thank you. And if a relaxing the standards is not going to happen then maybe some direction on how to make it fit the standard?
