### 2024-09-11 #657 @spsanderson
@mpadge I will take a look now. I think it should fit EDA at face value as there is plotting, parameter estimation, summary table functions and AIC calculation functions.
