### 2024-09-11 #657 @spsanderson
@mpadge

Here is an example of why I think it may fit with EDA as well:

## Plotting:
```r
library(TidyDensity)
library(dplyr)

# Create a data frame with 100 random numbers
set.seed(123)
df_tbl <- tidy_normal(.n = 100, .num_sims = 5)

# Plot different aspects of the data
# The default plot is a density plot
tidy_autoplot(df_tbl)
```
![image](https://github.com/user-attachments/assets/ec42e403-63d6-4f67-b251-e2f43b47dc7c)

```r
# A qq plot
tidy_autoplot(df_tbl, .plot_type = "qq")
```
![image](https://github.com/user-attachments/assets/831ef079-649f-4e51-8745-7404cbb4d9f2)

```r
# A quantile plot
tidy_autoplot(df_tbl, .plot_type = "quantile")
```
![image](https://github.com/user-attachments/assets/77566c73-5deb-4387-9d18-8a733163d85c)

```r
# A probability plot
tidy_autoplot(df_tbl, .plot_type = "probability")
```
![image](https://github.com/user-attachments/assets/a3678f40-1757-40c5-9c65-c0a62ac05791)

## Summary Stats:
```r
# A tibble: 1 × 12
  mean_val median_val std_val min_val max_val skewness kurtosis range   iqr variance ci_low ci_high
     <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl> <dbl>    <dbl>  <dbl>   <dbl>
1   0.0346     0.0207   0.973   -2.66    3.24   0.0861     2.95  5.90  1.26    0.946  -1.77    2.00
> tidy_distribution_summary_tbl(df_tbl, sim_number)
# A tibble: 5 × 13
  sim_number mean_val median_val std_val min_val max_val skewness kurtosis range   iqr variance
  <fct>         <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl> <dbl>    <dbl>
1 1            0.0904    0.0618    0.913   -2.31    2.19   0.0605     2.84  4.50  1.19    0.833
2 2           -0.108    -0.226     0.967   -2.05    3.24   0.639      3.66  5.29  1.27    0.935
3 3            0.120     0.0359    0.950   -1.76    2.29   0.326      2.46  4.05  1.29    0.902
4 4           -0.0362   -0.00351   1.04    -2.47    2.57  -0.0312     2.74  5.04  1.42    1.08
5 5            0.106     0.165     0.989   -2.66    2.40  -0.431      3.31  5.06  1.12    0.979
# ℹ 2 more variables: ci_low <dbl>, ci_high <dbl>
```

## Bootstraping the data:
```r
> # Bootstrap plotting a single simulation of df_tbl
> tidy_bootstrap(filter(df_tbl, sim_number == 1)[["y"]])
# A tibble: 2,000 × 2
   sim_number bootstrap_samples
   <fct>      <list>
 1 1          <dbl [80]>
 2 2          <dbl [80]>
 3 3          <dbl [80]>
 4 4          <dbl [80]>
 5 5          <dbl [80]>
 6 6          <dbl [80]>
 7 7          <dbl [80]>
 8 8          <dbl [80]>
 9 9          <dbl [80]>
10 10         <dbl [80]>
# ℹ 1,990 more rows
# ℹ Use `print(n = ...)` to see more rows
```

## Bootstrap Plots
```r
tidy_bootstrap(filter(df_tbl, sim_number == 1)[["y"]]) |>
  bootstrap_stat_plot(y)
```
![image](https://github.com/user-attachments/assets/a2060ba1-7e4c-4001-a537-2c9c3e0a3d6a)

```r
tidy_bootstrap(filter(df_tbl, sim_number == 1)[["y"]], .num_sims = 100) |>
  bootstrap_stat_plot(y, .show_groups = TRUE)
Warning message:
Setting '.num_sims' to less than 2000 means that results can be potentially unstable. Consider
setting to 2000 or more.
```
![image](https://github.com/user-attachments/assets/2bb90f9a-0f61-46b4-b05c-50cba800f620)

```r
> tidy_bootstrap(filter(df_tbl, sim_number == 1)[["y"]], .num_sims = 100) |>
+   bootstrap_stat_plot(y, .show_groups = TRUE, .stat = "cmax")
Warning message:
Setting '.num_sims' to less than 2000 means that results can be potentially unstable. Consider
setting to 2000 or more.
```
![image](https://github.com/user-attachments/assets/7ee26589-b05d-44e1-aa3c-c4f511663c0a)

## Stats about the distribution
```r
> # Stats about the distribution
> # All distributions
> util_normal_stats_tbl(df_tbl) |>
+   glimpse()
Rows: 1
Columns: 17
$ tidy_function     <chr> "tidy_gaussian"
$ function_call     <chr> "Gaussian c(0, 1)"
$ distribution      <chr> "Gaussian"
$ distribution_type <chr> "continuous"
$ points            <dbl> 100
$ simulations       <dbl> 5
$ mean              <dbl> 0
$ median            <dbl> 0.02071715
$ mode              <dbl> 0
$ std_dv            <dbl> 1
$ coeff_var         <dbl> Inf
$ skewness          <dbl> 0
$ kurtosis          <dbl> 3
$ computed_std_skew <dbl> 0.0861201
$ computed_std_kurt <dbl> 2.953595
$ ci_lo             <dbl> -1.774248
$ ci_hi             <dbl> 1.99998
> # A single distribution
> util_normal_stats_tbl(df_tbl |> filter(sim_number == 2)) |>
+   glimpse()
Rows: 1
Columns: 17
$ tidy_function     <chr> "tidy_gaussian"
$ function_call     <chr> "Gaussian c(0, 1)"
$ distribution      <chr> "Gaussian"
$ distribution_type <chr> "continuous"
$ points            <dbl> 100
$ simulations       <dbl> 5
$ mean              <dbl> 0
$ median            <dbl> -0.22583
$ mode              <dbl> 0
$ std_dv            <dbl> 1
$ coeff_var         <dbl> Inf
$ skewness          <dbl> 0
$ kurtosis          <dbl> 3
$ computed_std_skew <dbl> 0.6387938
$ computed_std_kurt <dbl> 3.656954
$ ci_lo             <dbl> -1.610118
$ ci_hi             <dbl> 2.051234
```

## Random walks of data and Viz
```r
# Create and plot a random walk of the data
tidy_random_walk(df_tbl |> filter(sim_number == 1), .initial_value = 100,
                 .value_type = "cum_prod") |>
  tidy_random_walk_autoplot()
```
![image](https://github.com/user-attachments/assets/8ed13846-5c1a-4bf4-83d0-b13a58f329cc)

```r
tidy_random_walk(df_tbl, .initial_value = 100, .value_type = "cum_sum") |>
  tidy_random_walk_autoplot()
```
![image](https://github.com/user-attachments/assets/d6654d2e-2809-4dd3-9617-41b80cfe462d)

## Getting AIC
```r
> # Getting AIC of distribution
> util_normal_aic(df_tbl |> filter(sim_number == 1) |> pull(y))
[1] 268.5385
```
