### 2024-10-15 #657 @spsanderson
@mpadge I just wanted to check back in and get thoughts of if I should also check off the EDA.

Thank you :)
