### 2024-10-23 #657 @mpadge
@spsanderson Yes, I definitely think so. That category is "Exploratory Data Analysis and Summary Statistics", and your package certainly has a strong focus on that latter component. Seems like you'd easily tick off at least half of those standards, so please go ahead and document compliance with those. When you're done you can call `@ropensci-review-bot check srr` yourself to get a statement of compliance and we'll proceed from there. Thanks!
