### 2024-10-24 #657 @mpadge
Yes, you should indeed. You'll need to document compliance with the statistical standards. The general procedure is described in https://stats-devguide.ropensci.org, and the specific tools for documenting compliance within your code as part of our [srr package](https://docs.ropensci.org/srr/articles/srr-stats.html). (The bot response above just indicates that you've not yet done that.) Feel free to ask any questions you might have during the process.
