### 2024-11-21 #657 @mpadge
@spsanderson That's getting there, but the idea is to write a statement of *how* you comply with each standard. A single sentence generally suffices. The compliance statements are intended to help reviews instantly see how the code they're about to look at complies with the standard. Have a look at [some of our other stats packages](https://github.com/ropensci/software-review/issues?q=is%3Aissue+label%3Astats+is%3Aclosed+label%3A6%2Fapproved-bronze-v0.1%2C6%2Fapproved-bronze-v0.2%2C6%2Fapproved-gold-v0.1%2C6%2Fapproved-gold-v0.2%2C6%2Fapproved-silver-v0.1%2C6%2Fapproved-silver-v0.2+) for examples.
