### 2024-11-21 #657 @spsanderson
@mpadge  thank you! I'll work on that and then ask you to look again before I proceed with the other ~150+ functions.
