### 2025-02-02 #657 @maurolepore
Dear @spsanderson as the new rotating EiC I'm reviewing the status of all open issues. I see you've discussed this pre-submission with @mpadge. Is there anything else you'd like to share before we can decide if this package is in scope and ready for a full submission?
