### 2025-02-04 #657 @spsanderson
It is not yet ready i have to tag everything, im way behind in development
over the passed few months

Steven P Sanderson II, MPH
Amazon Author Profile
<https://www.amazon.com/stores/Steven-P.-Sanderson/author/B0CZ157DHK?ref_=pe_2466670_811284380&isDramIntegrated=true&shoppingPortalEnabled=true>
Book on Lulu <http://goo.gl/lmrlFI>
Personal Site <http://www.spsanderson.com>

On Sun, Feb 2, 2025 at 12:02 PM Mauro Lepore ***@***.***>
wrote:

> Dear @spsanderson <https://github.com/spsanderson> as the new rotating
> EiC I'm reviewing the status of all open issues. I see you've discussed
> this pre-submission with @mpadge <https://github.com/mpadge>. Is there
> anything else you'd like to share before we can decide if this package is
> in scope and ready for a full submission?
>
