### 2025-02-05 #657 @maurolepore
@spsanderson for reporting the status, and for opening a pre-submission issue.

This issue seems to have accomplished it's main goal, i.e. to discuss what it would take for this package to be in scope for a full submission. Seeing that it's not ready yet, I will go ahead and close this issue so you don't feel rushed and we can focus on ongoing work.

Whenever you're ready, feel free to open a new pre-submission or full-submission, pointing to this one and summarizing how it then meets our scope requirements.

Thanks for considering rOpenSci as a home for your work. We look forward to seeing you here again.
